pyproject.toml
[project]
name = "sre-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # CDK
    "aws-cdk.aws-bedrock-agentcore-alpha==2.238.0a0",
    "aws-cdk-lib==2.238.0",
    "constructs==10.4.5",
    "cdk-ecr-deployment==4.1.0",
    # AgentCore
    "bedrock-agentcore==1.3.1",
    "strands-agents==1.27.0",
    "strands-agents-tools==0.2.20",
    "aws-opentelemetry-distro==0.15.0",
    # Shared
    "mcp==1.26.0",
    "boto3==1.42.52",
    "botocore==1.42.52",
    "httpx==0.28.1",
    "requests==2.32.5",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.3",  # Match aws-opentelemetry-distro requirement
    # Dev tools
    "ruff==0.9.1",
    "pytest>=8.0.0",
    "pytest-env>=1.1.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []

[tool.ruff]
# Target Python 3.12+
target-version = "py312"

# Include src and tests directories
include = ["src/**/*.py", "tests/**/*.py", "app.py"]

# Line length
line-length = 100

[tool.ruff.lint]
# Enable recommended rules plus additional checks
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

# Ignore specific rules if needed
ignore = [
    "E501",  # Line too long (handled by formatter)
]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with 4 spaces
indent-style = "space"
